api/requirements.txt
aiosqlite==0.21.0
alembic==1.16.3
asyncpg==0.30.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.14
fastapi-cli==0.0.7
greenlet==3.2.3
httptools==0.6.4
httpx>=0.27.0
Jinja2==3.1.6
pip==25.1.1
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
pytest-asyncio==1.0.0
pytest-postgresql==7.0.2
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.0.1
aioredis==2.0.1
setuptools==49.2.1
sqlmodel==0.0.24
testcontainers==4.10.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
passlib==1.7.4
pyjwt==2.10.1
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
aiofiles>=25.1.0
